Another crack at getting the sharpshooter to work
also added the enemies to the sample scene

+++ RAIL_GRIND_RUMBLE/Assets/Scripts/Kevin/IDamageable.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IDamageable 
 {
-    void TakeDamage(int Damage);
+    void TakeDamage(float Damage);
 
     Transform GetTransform();
 }
